- fixed SIFT_Align_LayerSet.java to handle global minimization   and emtpy layers and featureless tiles

+++ ini/trakem2/display/Displayable.java
@@ -1169,7 +1169,7 @@ public abstract class Displayable extends DBObject {
 		return pDst;
 	}
 
-	public Point2D.Double inverseTransformPoint(final int px, final int py) {
+	public Point2D.Double inverseTransformPoint(final double px, final double py) {
 		final Point2D.Double pSrc = new Point2D.Double(px, py);
 		if (this.at.isIdentity()) return pSrc;
 		final Point2D.Double pDst = new Point2D.Double();
